Move `Artifact` to the "shared" package
