+++ pkg/cmd/run/download/http.go
@@ -2,7 +2,6 @@ package download
 
 import (
 	"archive/zip"
-	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -10,63 +9,21 @@ import (
 	"os"
 
 	"github.com/cli/cli/api"
-	"github.com/cli/cli/internal/ghinstance"
 	"github.com/cli/cli/internal/ghrepo"
+	"github.com/cli/cli/pkg/cmd/run/shared"
 )
-
-type Artifact struct {
-	Name        string `json:"name"`
-	Size        uint32 `json:"size_in_bytes"`
-	DownloadURL string `json:"archive_download_url"`
-	Expired     bool   `json:"expired"`
-}
 
 type apiPlatform struct {
 	client *http.Client
 	repo   ghrepo.Interface
 }
 
-func (p *apiPlatform) List(runID string) ([]Artifact, error) {
-	return ListArtifacts(p.client, p.repo, runID)
+func (p *apiPlatform) List(runID string) ([]shared.Artifact, error) {
+	return shared.ListArtifacts(p.client, p.repo, runID)
 }
 
 func (p *apiPlatform) Download(url string, dir string) error {
 	return downloadArtifact(p.client, url, dir)
-}
-
-func ListArtifacts(httpClient *http.Client, repo ghrepo.Interface, runID string) ([]Artifact, error) {
-	perPage := 100
-	path := fmt.Sprintf("repos/%s/%s/actions/artifacts?per_page=%d", repo.RepoOwner(), repo.RepoName(), perPage)
-	if runID != "" {
-		path = fmt.Sprintf("repos/%s/%s/actions/runs/%s/artifacts?per_page=%d", repo.RepoOwner(), repo.RepoName(), runID, perPage)
-	}
-
-	req, err := http.NewRequest("GET", ghinstance.RESTPrefix(repo.RepoHost())+path, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode > 299 {
-		return nil, api.HandleHTTPError(resp)
-	}
-
-	var response struct {
-		TotalCount uint16 `json:"total_count"`
-		Artifacts  []Artifact
-	}
-
-	dec := json.NewDecoder(resp.Body)
-	if err := dec.Decode(&response); err != nil {
-		return response.Artifacts, fmt.Errorf("error parsing JSON: %w", err)
-	}
-
-	return response.Artifacts, nil
 }
 
 func downloadArtifact(httpClient *http.Client, url, destDir string) error {
